Changed impr_generate_similar_queries_no_model_reload() function name to generate_similar_queries_all_topics().
Added to impr_generate_similar_queries() function the max_n_query to set the max desired number of queries to generate from the provided one.
In the same function, implemented a heuristics to limit the number of candidates considered for each masked word (called top_k), so that the number of combinations is kept below max_combinations value.